Move JWT processing to middleware
