--- app/Http/Middleware/ExampleMiddleware.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Http\Middleware;
-
-use Closure;
-
-class ExampleMiddleware
-{
-    /**
-     * Handle an incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Closure  $next
-     * @return mixed
-     */
-    public function handle($request, Closure $next)
-    {
-        return $next($request);
-    }
-}
